+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/minerva/MinervaViewModel.java
@@ -3,7 +3,7 @@ package be.ugent.zeus.hydra.ui.main.minerva;
 import android.app.Application;
 import android.util.Pair;
 
-import be.ugent.zeus.hydra.data.models.minerva.Course;
+import be.ugent.zeus.hydra.domain.models.minerva.Course;
 import be.ugent.zeus.hydra.repository.data.BaseLiveData;
 import be.ugent.zeus.hydra.repository.requests.Result;
 import be.ugent.zeus.hydra.ui.common.RefreshViewModel;
@@ -13,14 +13,14 @@ import java.util.List;
 /**
  * @author Niko Strijbol
  */
-public class MinervaViewModel extends RefreshViewModel<List<Pair<Course, Integer>>> {
+public class MinervaViewModel extends RefreshViewModel<List<Pair<Course, Long>>> {
 
     public MinervaViewModel(Application application) {
         super(application);
     }
 
     @Override
-    protected BaseLiveData<Result<List<Pair<Course, Integer>>>> constructDataInstance() {
+    protected BaseLiveData<Result<List<Pair<Course, Long>>>> constructDataInstance() {
         return new CourseLiveData(getApplication());
     }
